refactor(fsmeta): consolidate real-cluster integration tests
