+++ fsmeta/integration/e2e_test.go
@@ -7,19 +7,11 @@ import (
 	"testing"
 	"time"
 
-	coordclient "github.com/feichai0017/NoKV/coordinator/client"
 	"github.com/feichai0017/NoKV/fsmeta"
 	fsmetaclient "github.com/feichai0017/NoKV/fsmeta/client"
-	fsmetaexec "github.com/feichai0017/NoKV/fsmeta/exec"
 	fsmetaserver "github.com/feichai0017/NoKV/fsmeta/server"
-	"github.com/feichai0017/NoKV/raftstore/client"
-	"github.com/feichai0017/NoKV/raftstore/migrate"
-	raftmode "github.com/feichai0017/NoKV/raftstore/mode"
-	storepkg "github.com/feichai0017/NoKV/raftstore/store"
-	"github.com/feichai0017/NoKV/raftstore/testcluster"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 )
 
 func TestFSMetadataClientServerOnRealCluster(t *testing.T) {
@@ -124,56 +116,4 @@ func openFSMetadataClient(t *testing.T, ctx context.Context, executor fsmetaserv
 		grpcServer.Stop()
 		_ = ln.Close()
 	}
-}
-
-func openRealClusterExecutor(t *testing.T, ctx context.Context) *fsmetaexec.Executor {
-	t.Helper()
-
-	coord := testcluster.StartCoordinator(t)
-	t.Cleanup(func() { coord.Close(t) })
-
-	seedDir := t.TempDir()
-	standalone := testcluster.OpenStandaloneDB(t, seedDir, nil)
-	require.NoError(t, standalone.Close())
-
-	const (
-		storeID  = uint64(1)
-		regionID = uint64(171)
-		peerID   = uint64(101)
-	)
-	_, err := migrate.Init(migrate.InitConfig{
-		WorkDir:  seedDir,
-		StoreID:  storeID,
-		RegionID: regionID,
-		PeerID:   peerID,
-	})
-	require.NoError(t, err)
-
-	node := testcluster.StartNodeWithConfig(t, storeID, seedDir, testcluster.NodeConfig{
-		AllowedModes:      []raftmode.Mode{raftmode.ModeSeeded, raftmode.ModeCluster},
-		StartPeers:        true,
-		Scheduler:         testcluster.NewScheduler(t, coord.Addr(), 100*time.Millisecond),
-		HeartbeatInterval: 50 * time.Millisecond,
-	})
-	t.Cleanup(func() { node.Close(t) })
-
-	testcluster.WaitForLeaderPeer(t, ctx, node.Addr(), regionID, peerID)
-	testcluster.WaitForSchedulerMode(t, node, storepkg.SchedulerModeHealthy, false)
-
-	coordRPC, err := coordclient.NewGRPCClient(ctx, coord.Addr(), grpc.WithTransportCredentials(insecure.NewCredentials()))
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = coordRPC.Close() })
-	kv, err := client.New(client.Config{
-		StoreResolver:  coordRPC,
-		RegionResolver: coordRPC,
-		DialOptions:    []grpc.DialOption{grpc.WithTransportCredentials(insecure.NewCredentials())},
-	})
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = kv.Close() })
-
-	runner, err := fsmetaexec.NewRaftstoreRunner(kv, coordRPC)
-	require.NoError(t, err)
-	executor, err := fsmetaexec.New(runner)
-	require.NoError(t, err)
-	return executor
 }
